Fixed a bug in get_instance_segmentation that was causing invalid Polygons to raise an error.
